Entidades client, employee y company agregadas y relacionadas con address
