+++ src/main/java/com/example/buensabor/Address/City.java
@@ -1,0 +1,23 @@
+package com.example.buensabor.Address;
+
+import com.example.buensabor.Bases.BaseEntity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class City extends BaseEntity {
+
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "province_id")
+    private Province province;
+}
