+++ src/main/java/com/example/buensabor/Address/Country.java
@@ -1,0 +1,16 @@
+package com.example.buensabor.Address;
+
+import com.example.buensabor.Bases.BaseEntity;
+
+import jakarta.persistence.Entity;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Country extends BaseEntity{
+    private String name;
+}
