+++ src/main/java/com/example/buensabor/Client/ClientAddress.java
@@ -1,0 +1,28 @@
+package com.example.buensabor.Client;
+
+import com.example.buensabor.Address.Address;
+import com.example.buensabor.Bases.BaseEntity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "clients_addresses")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ClientAddress extends BaseEntity{
+
+    @ManyToOne
+    @JoinColumn(name = "client_id", nullable = false)
+    private Client client;
+
+    @ManyToOne
+    @JoinColumn(name = "address_id", nullable = false)
+    private Address address;
+}
