+++ src/main/java/com/example/buensabor/Company/Company.java
@@ -1,0 +1,18 @@
+package com.example.buensabor.Company;
+
+import com.example.buensabor.Address.Address;
+import com.example.buensabor.User.User;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "companies")
+public class Company extends User{
+    
+    @OneToOne
+    @JoinColumn(name = "address_id", nullable = false)
+    private Address address;
+}
