+++ src/main/java/com/example/buensabor/Ingredient/IngredientRepository.java
@@ -1,0 +1,10 @@
+package com.example.buensabor.Ingredient;
+
+import org.springframework.stereotype.Repository;
+
+import com.example.buensabor.Bases.BaseRepository;
+
+@Repository
+public interface IngredientRepository extends BaseRepository<Ingredient, Long> {
+    
+}
