+++ src/main/java/com/example/buensabor/Ingredient/Interfaces/IIngredientController.java
@@ -1,0 +1,8 @@
+package com.example.buensabor.Ingredient.Interfaces;
+
+import com.example.buensabor.Bases.IBaseController;
+import com.example.buensabor.Ingredient.Ingredient;
+
+public interface IIngredientController extends IBaseController<Ingredient, Long> {
+    
+}
