+++ src/main/java/com/example/buensabor/Ingredient/Interfaces/IIngredientService.java
@@ -1,0 +1,8 @@
+package com.example.buensabor.Ingredient.Interfaces;
+
+import com.example.buensabor.Bases.IBaseService;
+import com.example.buensabor.Ingredient.Ingredient;
+
+public interface IIngredientService extends IBaseService<Ingredient, Long> {
+    
+}
