+++ src/main/java/com/example/buensabor/Product/Interfaces/IProductController.java
@@ -1,0 +1,8 @@
+package com.example.buensabor.Product.Interfaces;
+
+import com.example.buensabor.Bases.IBaseController;
+import com.example.buensabor.Product.Product;
+
+public interface IProductController extends IBaseController<Product, Long> {
+    
+}
