+++ src/main/java/com/example/buensabor/Product/Interfaces/IProductService.java
@@ -1,0 +1,8 @@
+package com.example.buensabor.Product.Interfaces;
+
+import com.example.buensabor.Bases.IBaseService;
+import com.example.buensabor.Product.Product;
+
+public interface IProductService extends IBaseService<Product, Long> {
+    
+}
