+++ src/main/java/com/example/buensabor/Product/ProductController.java
@@ -1,0 +1,15 @@
+package com.example.buensabor.Product;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.buensabor.Bases.BaseControllerImplementation;
+import com.example.buensabor.Product.Interfaces.IProductController;
+
+@RestController
+@CrossOrigin(origins = "*")
+@RequestMapping(path = "api/v1/products")
+public class ProductController extends BaseControllerImplementation<Product, ProductService> implements IProductController {
+    // ...existing code...
+}
