+++ src/main/java/com/example/buensabor/Product/ProductRepository.java
@@ -1,0 +1,10 @@
+package com.example.buensabor.Product;
+
+import org.springframework.stereotype.Repository;
+
+import com.example.buensabor.Bases.BaseRepository;
+
+@Repository
+public interface ProductRepository extends BaseRepository<Product, Long> {
+    
+}
