+++ src/main/java/com/example/buensabor/ProductIngredient/ProductIngredientRepository.java
@@ -1,0 +1,10 @@
+package com.example.buensabor.ProductIngredient;
+
+import org.springframework.stereotype.Repository;
+
+import com.example.buensabor.Bases.BaseRepository;
+
+@Repository
+public interface ProductIngredientRepository extends BaseRepository<ProductIngredient, Long> {
+    
+}
