--- src/main/java/com/example/buensabor/Products/Ingredient/IngredientRepository.java
@@ -1,10 +1,0 @@
-package com.example.buensabor.Products.Ingredient;
-
-import org.springframework.stereotype.Repository;
-
-import com.example.buensabor.Bases.BaseRepository;
-
-@Repository
-public interface IngredientRepository extends BaseRepository<Ingredient, Long> {
-    
-}
