--- src/main/java/com/example/buensabor/Products/Ingredient/Interfaces/IIngredientController.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.Products.Ingredient.Interfaces;
-
-import com.example.buensabor.Bases.IBaseController;
-import com.example.buensabor.Products.Ingredient.Ingredient;
-
-public interface IIngredientController extends IBaseController<Ingredient, Long> {
-    
-}
