--- src/main/java/com/example/buensabor/Products/Ingredient/Interfaces/IIngredientService.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.Products.Ingredient.Interfaces;
-
-import com.example.buensabor.Bases.IBaseService;
-import com.example.buensabor.Products.Ingredient.Ingredient;
-
-public interface IIngredientService extends IBaseService<Ingredient, Long> {
-    
-}
