--- src/main/java/com/example/buensabor/Products/Product/Interfaces/IProductController.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.Products.Product.Interfaces;
-
-import com.example.buensabor.Bases.IBaseController;
-import com.example.buensabor.Products.Product.Product;
-
-public interface IProductController extends IBaseController<Product, Long> {
-    
-}
