--- src/main/java/com/example/buensabor/Products/Product/Interfaces/IProductService.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.Products.Product.Interfaces;
-
-import com.example.buensabor.Bases.IBaseService;
-import com.example.buensabor.Products.Product.Product;
-
-public interface IProductService extends IBaseService<Product, Long> {
-    
-}
