--- src/main/java/com/example/buensabor/Products/Product/ProductRepository.java
@@ -1,10 +1,0 @@
-package com.example.buensabor.Products.Product;
-
-import org.springframework.stereotype.Repository;
-
-import com.example.buensabor.Bases.BaseRepository;
-
-@Repository
-public interface ProductRepository extends BaseRepository<Product, Long> {
-    
-}
