--- src/main/java/com/example/buensabor/Products/ProductIngredient/ProductIngredientRepository.java
@@ -1,10 +1,0 @@
-package com.example.buensabor.Products.ProductIngredient;
-
-import org.springframework.stereotype.Repository;
-
-import com.example.buensabor.Bases.BaseRepository;
-
-@Repository
-public interface ProductIngredientRepository extends BaseRepository<ProductIngredient, Long> {
-    
-}
