--- src/main/java/com/example/buensabor/User/Interfaces/IUserController.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.User.Interfaces;
-
-import com.example.buensabor.Bases.IBaseController;
-import com.example.buensabor.User.User;
-
-public interface IUserController extends IBaseController<User, Long> {
-    
-}
