--- src/main/java/com/example/buensabor/User/Interfaces/IUserService.java
@@ -1,8 +1,0 @@
-package com.example.buensabor.User.Interfaces;
-
-import com.example.buensabor.Bases.IBaseService;
-import com.example.buensabor.User.User;
-
-public interface IUserService extends IBaseService<User, Long> {
-    
-}
