+++ src/main/java/com/example/buensabor/User/User.java
@@ -7,19 +7,11 @@ import com.example.buensabor.Bases.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
+import jakarta.persistence.Inheritance;
+import jakarta.persistence.InheritanceType;
 @Entity
-@Table(name = "users")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class User extends BaseEntity {
+@Inheritance(strategy = InheritanceType.JOINED)  // o SINGLE_TABLE, TABLE_PER_CLASS
+public abstract class User extends BaseEntity {
     
     @Column(name = "name")
     private String name;
