--- src/main/java/com/example/buensabor/User/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.buensabor.User;
-
-import org.springframework.stereotype.Repository;
-
-import com.example.buensabor.Bases.BaseRepository;
-
-@Repository
-public interface UserRepository extends BaseRepository<User, Long> {
-    
-}
