+++ src/main/java/com/example/buensabor/enums/RoleEmployee.java
@@ -1,0 +1,7 @@
+package com.example.buensabor.enums;
+
+public enum RoleEmployee {
+    CASHIER,
+    COOK,
+    DELIVERY
+}
